pyproject.toml
[tool.poetry]
name = "firecrawl_tool"
version = "0.1.0"
description = "A tool to scrape or extract data from a website using FireCrawl"
authors = ["Dean <dean@aidevmode.com>"]
readme = "README.md"
packages = [
    { include = "firecrawl_tool", from = "." }
]

[tool.poetry.dependencies]
python = ">=3.10,<=3.13"
naptha-sdk = {git = "https://github.com/NapthaAI/naptha-sdk.git"}
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
